BEAM SEARCH
Beam search is a popular search algorithm used in natural language processing and machine translation tasks. It is employed to find the most likely sequence of words, given a sequence of input words. Beam search is particularly useful in tasks such as machine translation, text summarization, and speech recognition, where the goal is to generate a sequence of words that maximizes the probability of the sequence being correct and coherent.